Replace BottomSheet.SubSheet with individual React Navigation screens
The current `BottomSheet.SubSheet` implementation does not reset a given
sub sheets visibility when navigating backwards via the Android back
button. This results in the child content remaining visible. When
subsequently navigating to a new child, the previous child is also
displayed.

We might be able to hook into React Navigation events to resolve this,
but my exploration felt forced and not all that valuable.
